refactor: extract GitHub event formatters to separate module
Extract all 7 GitHub webhook event formatters from monolithic index.ts:

- Create src/formatters/github-events.ts module
  - formatPullRequest (opened, merged, closed)
  - formatIssue (opened, closed)
  - formatPush (commits with branch info)
  - formatRelease (published)
  - formatWorkflowRun (CI pass/fail)
  - formatIssueComment (new comments)
  - formatPullRequestReview (approved, changes requested)

- Remove ~200 lines from src/index.ts
- Import formatters from centralized module
- Maintains all existing functionality

Benefits:
- Better separation of concerns
- Easier to test formatters in isolation
- Reusable across different contexts
- Smaller, more focused files

All 33 tests still passing ✓
